test dedup: consolidate execution-layer protocol and result-builder tests
- Unify _snap() in conftest: merge two incompatible signatures, add cpu_percent
  and full signal/thread/ctx_switch parameterization. Use partial with
  captured_at=None for anomaly detection tests that omit timestamps.
- Centralize _result_ndjson() and _assistant_ndjson() in conftest: add to
  conftest, remove duplicate local definitions from test_session_result.py
  (which never called them) and test_session_parsing.py.
- Migrate test_process_pty.py to conftest._result_ndjson: replace local
  definition, update 3 call sites from result= to result_text=.
- Delete C2-4 redundant protocol tests: test_github_fetcher_protocol_includes_write_methods
  and TestFetchTitle.test_protocol_conformance (superseded by
  test_default_github_fetcher_implements_full_protocol).
- Delete C2-9 duplicate null-coercion tests:
  test_post_init_coerces_none_result_to_empty_str and
  test_post_init_coerces_non_list_errors (retained unique test
  test_post_init_list_with_non_dict_element).
